Now almost a complete metrics stack
I'm using Haggar running on host os to drive metrics
Not sure why I can not use port 2003 for otel-in :-/
Most importantly, export from Redpanda to carbon-relay-ng is shockly
slow. Maybe as low as 100 events a second. I thought this might be due
to a bad interaction between otel and carbon-clickhouse so I shimmed in
carbon-relay-ng but the performance is about the same..